Comark.Parser.Options

Documentation

data ParserOption

Constructors

Normalize
Consolidate adjacent text nodes.

LinkReferences (Text -> Maybe (Text, Maybe Text))
Predefine link reference defenitions.

References are represented with a mapping from a link text to a pair of a link destination and an optional link title.

During parsing the link references defined in a document would be collected into additional mapping. When link references are being mapping defined in options takes precedence over mapping found in the document.

TODO: Examples

data ParserOptions

parserOptions :: [ParserOption] -> ParserOptions

_poNormalize :: ParserOptions -> Bool

_poLinkReferences :: ParserOptions -> Text -> Maybe (Text, Maybe Text)

_poParseEmphasis :: ParserOptions -> Bool