Comark.ParserCombinators

Documentation

data Position

Constructors

Position

Fields

line :: Int

column :: Int

point :: Int

data Parser a

data ParseError

Constructors

ParseError

Fields

parseErrorPosition :: Position

parseErrorReason :: String

runParser :: Parser a -> Text -> Either ParseError a

runParserWithUnconsumed :: Parser a -> Text -> Either ParseError (a, Text)

(<?>) :: Parser a -> String -> Parser a infixl 5

satisfy :: (Char -> Bool) -> Parser Char

withConsumed :: Parser a -> Parser (a, Text)

Returns the text that was consumed by a parser alongside with its result

consumedBy :: Parser a -> Parser Text

peekChar :: Parser (Maybe Char)

Get the next character without consuming it.

peekLastChar :: Parser (Maybe Char)

Get the last consumed character.

notAfter :: (Char -> Bool) -> Parser ()

inClass :: String -> Char -> Bool

notInClass :: String -> Char -> Bool

endOfInput :: Parser ()

char :: Char -> Parser Char

anyChar :: Parser Char

getPosition :: Parser Position

setPosition :: Position -> Parser ()

takeWhile :: (Char -> Bool) -> Parser Text

takeTill :: (Char -> Bool) -> Parser Text

takeWhile1 :: (Char -> Bool) -> Parser Text

untilTheEnd :: Parser Text

Consumes all the available input (until endOfInput) and returns it.

skip :: (Char -> Bool) -> Parser ()

skipWhile :: (Char -> Bool) -> Parser ()

skipWhile1 :: (Char -> Bool) -> Parser ()

replacing :: Parser Text -> Parser ()

Takes a parser that returns a Text, runs that parser consuming some input and then prepends the result to the rest of subject.

string :: Text -> Parser Text

stringCaseless :: Text -> Parser Text

scan :: s -> (s -> Char -> Maybe s) -> Parser Text

lookAhead :: Parser a -> Parser a

notFollowedBy :: Parser a -> Parser ()

option :: Alternative f => a -> f a -> f a

foldP :: (b -> Parser (Maybe b)) -> b -> Parser b

manyTill :: Alternative f => f a -> f b -> f [a]

someTill :: Alternative f => f a -> f b -> f [a]

sepBy1 :: Alternative f => f a -> f s -> f [a]

sepEndBy1 :: Alternative f => f a -> f s -> f [a]

sepStartEndBy1 :: Alternative f => f a -> f s -> f [a]

skipMany :: Alternative f => f a -> f ()

skipMany1 :: Alternative f => f a -> f ()

discardOpt :: Alternative f => f a -> f ()

decimal :: Integral a => Parser a

Parse and decode an unsigned decimal number.

hexadecimal :: (Integral a, Bits a) => Parser a