Data.Text.Extended

Documentation

module Data.Text

tabFilter :: Int -> Text -> Text

joinLines :: [Text] -> Text

Like unlines but does not add a final newline. | Concatenates lines with newlines between.

lines' :: Text -> [Text]

Been commented out from Data.Text O(n) Portably breaks a Text up into a list of Texts at line boundaries.

A line boundary is considered to be either a line feed, a carriage return immediately followed by a line feed, or a carriage return. This accounts for both Unix and Windows line ending conventions, and for the old convention used on Mac OS 9 and earlier.